What is the board?

This grid contains all the categories, clues, and answers in a game.

whatr_board(game)

Format

A tibble with (usually) 61 rows and 4 variables:

round

The round a clue is chosen.

col

The column position left-to-right.

row

The row position top-to-bottom.

n

The order of clue chosen.

category

Category title, often humorous or with instructions.

clue

The clue read to the contestants.

answer

The correct answer to a clue.

Arguments

game

One of four types of input, all passed to whatr_html():

A numeric game ID.

Either a 'showgame' or 'showscores' HTML document.

A show number character starting with "#".

An air date like "yyyy-mm-dd".

Value

A tidy tibble of clue text.

Examples

whatr_board(game = 6304)
#> # A tibble: 61 × 7
#>    round   col   row     i category          clue                         answer
#>    <int> <int> <int> <int> <chr>             <chr>                        <chr> 
#>  1     1     1     1    25 Picture The Idiom For An Optimistic View Of T… Rose-…
#>  2     1     1     2    24 Picture The Idiom If You Have These, It Means… Your …
#>  3     1     1     3     7 Picture The Idiom It's What's Going On Here    Compa…
#>  4     1     1     4     6 Picture The Idiom Greenhouse Gas Emission Due… A Car…
#>  5     1     1     5     4 Picture The Idiom Note The Lovely Weather; It… Make …
#>  6     1     2     1    26 Literature        Melville's 'Bartleby The Sc… Wall …
#>  7     1     2     2    23 Literature        In A 1923 Book By Kahlil Gi… The P…
#>  8     1     2     3    15 Literature        In 'Charlotte's Web', Templ… A Rat 
#>  9     1     2     4     3 Literature        In A Novel Simone De Beauvo… (Albe…
#> 10     1     2     5     1 Literature        The Title Peak Of This Thom… Magic…
#> # … with 51 more rows