Scrape Jeopardy game clues

Confusingly, clues are "questions" posed to contestants in the form of an answer. The clues are the questions read by the host which must be answered by the contestants.

whatr_clues(game = NULL, date = NULL, show = NULL)

Arguments

game
The J-Archive! game ID number.

date
The original date an episode aired.

show
The sequential show number.

Value

A tidy tibble or correct answers.

Details

From the J! Archive glossary: [Clues are] a "question" posed to Jeopardy! contestants, descriptively phrased as the answer to the correct response. Clues have some dollar value associated with them, except in cases of Daily Doubles or Final Jeopardy! clues, which are wagered upon prior to the revealing of the clue... At the start of the Jeopardy! Round, the returning champion, or the player at the leftmost lectern if there is no returning champion, selects the first clue; thereafter, for the rest of the round, the clues are selected by the last player to give a correct response. At the beginning of the Double Jeopardy! Round, the player with the lowest score selects the first clue.

Examples

whatr_clues(game = 6304)

#> # A tibble: 61 x 5
#>    round   col   row     n clue                                                 
#>    <int> <dbl> <dbl> <int> <chr>                                                
#>  1     1     1     1    25 For An Optimistic View Of Things, Take A Look Throug…
#>  2     1     2     1    26 Melville's 'Bartleby The Scrivener' Is Subtitled 'A …
#>  3     1     3     1    27 Ahoy!  It's A Lookout Perch Atop A High Mast         
#>  4     1     4     1    28 'Escape To Margaritaville':Him                       
#>  5     1     5     1    29 'The American People Don't Believe Anything's Real U…
#>  6     1     6     1    30 I'm Not A Cook, But I Can Sure Toast Up These Kellog…
#>  7     1     1     2    24 If You Have These, It Means You're Well-Organized    
#>  8     1     2     2    23 In A 1923 Book By Kahlil Gibran, Almustafa Is This M…
#>  9     1     3     2    22 Made In A Series Of Varying Sizes To Fit One Inside …
#> 10     1     4     2    21 'Jersey Boys':Them                                   
#> # … with 51 more rows