What is the board?

This is the name for game mechanic displaying all categories and clues.

whatr_data(game)

Arguments

game
One of four types of input, all passed to whatr_html():

A numeric game ID.

Either a 'showgame' or 'showscores' HTML document.

A show number character starting with "#".

An air date like "yyyy-mm-dd".

Value

A list of tibbles.

Examples

whatr_data(game = 6304)

#> $info
#> # A tibble: 1 x 3
#>    game  show date      
#>   <int> <int> <date>    
#> 1  6304  8006 2019-06-03
#> 
#> $summary
#> # A tibble: 3 x 5
#>   name  final coryat right wrong
#>   <chr> <int>  <int> <int> <int>
#> 1 James 24799  23400    25     0
#> 2 Jay   17000  11000    13     1
#> 3 Emma  46801  18800    21     0
#> 
#> $players
#> # A tibble: 3 x 4
#>   first last      occupation                  from             
#>   <chr> <chr>     <chr>                       <chr>            
#> 1 Emma  Boettcher User Experience Librarian   Chicago, Illinois
#> 2 Jay   Sexton    Senior Research Engineer    Atlanta, Georgia 
#> 3 James Holzhauer Professional Sports Gambler Las Vegas, Nevada
#> 
#> $scores
#> # A tibble: 63 x 5
#>    round     i name  score double
#>    <int> <int> <chr> <int> <lgl> 
#>  1     1     1 James  1000 TRUE  
#>  2     1     2 Emma   1000 FALSE 
#>  3     1     3 James   800 FALSE 
#>  4     1     4 James  1000 FALSE 
#>  5     1     5 Emma   1000 FALSE 
#>  6     1     6 Jay     800 FALSE 
#>  7     1     7 James   600 FALSE 
#>  8     1     8 James  1000 FALSE 
#>  9     1     9 James  1000 FALSE 
#> 10     1    10 Emma    800 FALSE 
#> # … with 53 more rows
#> 
#> $board
#> # A tibble: 61 x 7
#>    round   col   row     i category    clue                        answer       
#>    <int> <int> <int> <int> <chr>       <chr>                       <chr>        
#>  1     1     1     1    25 Picture Th… For An Optimistic View Of … Rose-Colored…
#>  2     1     1     2    24 Picture Th… If You Have These, It Mean… Your Ducks I…
#>  3     1     1     3     7 Picture Th… It's What's Going On Here   Comparing Ap…
#>  4     1     1     4     6 Picture Th… Greenhouse Gas Emission Du… A Carbon Foo…
#>  5     1     1     5     4 Picture Th… Note The Lovely Weather; I… Make Hay Whi…
#>  6     1     2     1    26 Literature  Melville's 'Bartleby The S… Wall Street  
#>  7     1     2     2    23 Literature  In A 1923 Book By Kahlil G… The Prophet  
#>  8     1     2     3    15 Literature  In 'Charlotte's Web', Temp… A Rat        
#>  9     1     2     4     3 Literature  In A Novel Simone De Beauv… (Albert) Cam…
#> 10     1     2     5     1 Literature  The Title Peak Of This Tho… Magic Mounta…
#> # … with 51 more rows
#>