Return the right J! Archive HTML document

To avoid downloading and reading the J! Archive over and over, this function can be used to return an HTML document from one of four types of input:

Game ID (6304)

Show number ("#8006")

Air date ("2019-06-03")

HTML document itself

whatr_html(x, out = c("showgame", "showscores"))

Arguments

x
Any kind of J! Archive game identifier.

out
One of "showscores" or "showgame" used for whatr_scores() or whatr_clues(), etc. respectively. Either document can be returned from the input of the other.

Value

A J! Archive html_document.

Examples

whatr_html(x = 6304, out = "showscores")

#> {html_document}
#> <html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en">
#> [1] <head>\n<meta http-equiv="Content-Type" content="text/html; charset=UTF-8 ...
#> [2] <body>\n\n<div id="navbar">\n  <span id="navbarlogo"><a href="http://www. ...

whatr_html(x = "2019-06-03", out = "showgame")

#> {html_document}
#> <html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en">
#> [1] <head>\n<meta http-equiv="Content-Type" content="text/html; charset=UTF-8 ...
#> [2] <body>\n\n<div id="navbar">\n  <span id="navbarlogo"><a href="http://www. ...

whatr_html("#8006", "showgame") %>% whatr_clues()

#> # A tibble: 61 x 5
#>    round   col   row     n clue                                                 
#>    <int> <int> <int> <int> <chr>                                                
#>  1     1     1     1    25 For An Optimistic View Of Things, Take A Look Throug…
#>  2     1     2     1    26 Melville's 'Bartleby The Scrivener' Is Subtitled 'A …
#>  3     1     3     1    27 Ahoy!  It's A Lookout Perch Atop A High Mast         
#>  4     1     4     1    28 'Escape To Margaritaville':Him                       
#>  5     1     5     1    29 'The American People Don't Believe Anything's Real U…
#>  6     1     6     1    30 I'm Not A Cook, But I Can Sure Toast Up These Kellog…
#>  7     1     1     2    24 If You Have These, It Means You're Well-Organized    
#>  8     1     2     2    23 In A 1923 Book By Kahlil Gibran, Almustafa Is This M…
#>  9     1     3     2    22 Made In A Series Of Varying Sizes To Fit One Inside …
#> 10     1     4     2    21 'Jersey Boys':Them                                   
#> # … with 51 more rows