Return all a game's data from the J! Archive

Return all a game's data from the J! Archive

whatr_data(game = NULL, date = NULL, show = NULL)

Arguments

game
The J-Archive! game ID number.

date
The original date an episode aired.

show
The sequential show number.

Value

A list of tibbles each containing a data type.

Examples

whatr_data(game = 6304)

#> $info
#> # A tibble: 1 x 3
#>    game  show date      
#>   <int> <int> <date>    
#> 1  6304  8006 2019-06-03
#> 
#> $players
#> # A tibble: 3 x 5
#>   first last      occupation                  city      state   
#>   <chr> <chr>     <chr>                       <chr>     <chr>   
#> 1 Emma  Boettcher User Experience Librarian   Chicago   Illinois
#> 2 Jay   Sexton    Senior Research Engineer    Atlanta   Georgia 
#> 3 James Holzhauer Professional Sports Gambler Las Vegas Nevada  
#> 
#> $order
#> # A tibble: 61 x 4
#>    round   col   row     n
#>    <int> <int> <int> <int>
#>  1     1     2     5     1
#>  2     1     3     5     2
#>  3     1     2     4     3
#>  4     1     1     5     4
#>  5     1     4     5     5
#>  6     1     1     4     6
#>  7     1     1     3     7
#>  8     1     5     5     8
#>  9     1     6     5     9
#> 10     1     6     4    10
#> # … with 51 more rows
#> 
#> $board
#> # A tibble: 61 x 4
#>        n category      clue                                    answer           
#>    <int> <chr>         <chr>                                   <chr>            
#>  1     1 Literature    The Title Peak Of This Thomas Mann Nov… Magic Mountain   
#>  2     2 Nests         Doesn't Sound Comfy, But Several Birds… A Woodpecker     
#>  3     3 Literature    In A Novel Simone De Beauvoir Depicted… (Albert) Camus   
#>  4     4 Picture The … Note The Lovely Weather; It Means To T… Make Hay While T…
#>  5     5 Jukebox Musi… 'Ain't Too Proud':This Motown Group     The Temptations  
#>  6     6 Picture The … Greenhouse Gas Emission Due To Human A… A Carbon Footpri…
#>  7     7 Picture The … It's What's Going On Here               Comparing Apples…
#>  8     8 Nixon Said It In His 1977 Interviews With This Briti… (David) Frost    
#>  9     9 I Am Not A C… You Don't Want Vegetables Cooked By Me… Crudités         
#> 10    10 I Am Not A C… I'm A Wizard At Breakfast With This Ce… Lucky Charms     
#> # … with 51 more rows
#> 
#> $doubles
#> # A tibble: 3 x 4
#>       n name  correct change
#>   <int> <chr> <lgl>    <int>
#> 1     1 James TRUE      1000
#> 2    34 Emma  TRUE      7600
#> 3    44 Emma  TRUE      3000
#> 
#> $scores
#> # A tibble: 183 x 7
#>    round     n name  score double change correct
#>    <int> <int> <chr> <int> <lgl>   <int> <lgl>  
#>  1     1     1 James  1000 TRUE     1000 TRUE   
#>  2     1     1 Jay       0 TRUE        0 NA     
#>  3     1     1 Emma      0 TRUE        0 NA     
#>  4     1     2 James  1000 FALSE       0 NA     
#>  5     1     2 Jay       0 FALSE       0 NA     
#>  6     1     2 Emma   1000 FALSE    1000 TRUE   
#>  7     1     3 James  1800 FALSE     800 TRUE   
#>  8     1     3 Jay       0 FALSE       0 NA     
#>  9     1     3 Emma   1000 FALSE       0 NA     
#> 10     1     4 James  2800 FALSE    1000 TRUE   
#> # … with 173 more rows
#> 
#> $summary
#> # A tibble: 3 x 5
#>   name  coryat final right wrong
#>   <chr>  <int> <int> <int> <int>
#> 1 James  23400 24799    25     0
#> 2 Jay    11000 17000    13     1
#> 3 Emma   18800 46801    21     0
#>